lib/vfscore: Use `unsigned long` for `fcntl`'s third argument's type
Although `fcntl` is most frequently used in conjunction with `int`s
as its third argument, the use case involving file locking with
operations such as `F_SETLK` or `F_GETLK` may require that the
third argument is a pointer.

Therefore, allow for correct addresses above 2G to be passed through
`fcntl` by changing the type of the third argument to `unsigned long`.

+++ lib/vfscore/include/vfscore/syscalls.h
@@ -30,7 +30,7 @@ int vfscore_lseek(struct vfscore_file *fp, off_t off, int type, off_t *origin);
 
 int vfscore_fstat(struct vfscore_file *fp, struct stat *st);
 
-int vfscore_fcntl(struct vfscore_file *fp, unsigned int cmd, int arg);
+int vfscore_fcntl(struct vfscore_file *fp, unsigned int cmd, unsigned long arg);
 int vfscore_ioctl(struct vfscore_file *fp, unsigned long request, void *buf);
 
 #endif /* __VFSCORE_SYSCALLS_H__ */
